Update the Model_Betasheet_AmideI with the new GUI pattern.
Make the OneExH property of the StrutureData class as a calculated property, which depends on the LocFreq and the Beta properties.
In this way, the modification of system coupling will be auto propagated to all the spectral calculations.

Also define all the dynamical properties as NonCopyable so when using the SD_Copy method, it won't trigger error even if there are not enough defined properties for instantiation.